make poetry package
