+++ pre2k/lib/commands/__init__.py
@@ -1,0 +1,1 @@
+from pre2k.lib.commands import auth, unauth
