--- requirements.txt
@@ -1,4 +1,0 @@
-impacket==0.10.0
-ldap3==2.9.1
-pyasn1==0.4.8
-rich==12.5.1
